// src/enum/access.enum.ts
export enum AccessEnum {
    DASHBOARD = 'Dashboard',
    USERS = 'Users',
    TRADES = 'Trades',
    VERIFICATION = 'Verification',
    WITHDRAWAL = 'Withdrawal',
    TRANSACTION = 'Transaction',
    CHAT = 'Chat',
    NOTIFICATION = 'Notification',
    EMAIL = 'Email',
    ADMINUSERS = 'AdminUsers',
    CONFIGURATION = 'Configuration'
}